+++ server/index.js
@@ -145,90 +145,6 @@ const db = mysql.createPool({
 
 app.use(require("./app/routes"));
 
-//Register
-// app.post("/register", (req, res) => {
-//   const firstName = req.body.firstName;
-
-//   const lastName = req.body.lastName;
-//   const email = req.body.email;
-//   const password = req.body.password;
-//   const userName = req.body.userName
-//   const date = new Date();
-
-//   const theme = "Dark";
-//   const timezone = "UTC+0";
-//   const currency = "GBP";
-//   const accountLevel = "Standard";
-
-//   //hash password
-//   bcrypt.hash(password, saltRounds, (err, hash) => {
-//     if (err) {
-//       console.log(err);
-//       res.send({ err });
-//     } else {
-//       db.query(
-//         "INSERT INTO users (firstName, lastName, email, password, userName, registeredDate) VALUES (?,?,?,?,?,?)",
-//         [firstName, lastName, email, hash, userName, date],
-//         (err, result) => {
-//           console.log(err);
-//         }
-//       );
-//       db.query(
-//         "INSERT INTO userSettings (theme, timezone, currency) VALUES (?,?,?)",
-//         [theme, timezone, currency],
-//         (err, result) => {
-//           console.log(err);
-//         }
-//       );
-//       db.query(
-//         "INSERT INTO accountLevel (accountLevel, dateUpgraded) VALUES (?,?)",
-//         [accountLevel, date,],
-//         (err, result) => {
-//           console.log(err);
-//         }
-//       );
-//       db.query(
-//         "INSERT INTO userAuth (Email, emailVerified, SMS, google, googleSecret, Authy, phoneNumber) VALUES (?,?,?,?,?,?,?)",
-//         [email, 0, 0, 0, 0, 0, "0"],
-//         (err, result) => {
-//           console.log(err);
-//         }
-//       );
-//       db.query(
-//         "INSERT INTO userPaymentAccounts (EUBank, UKBank, InterBank, Paypal, Skrill) VALUES (?,?,?,?,?)",
-//         [0, 0, 0, 0, 0],
-//         (err, result) => {
-//           console.log(err);
-//         }
-//       );
-//       res.send({
-//         registered: true,
-//       });
-//     }
-//   });
-// });
-
-app.post("/getUserInfo", (req, res) => {
-  const id = req.session.user[0].userID;
-  db.query("SELECT * FROM userInfo WHERE", [], (err, result) => {});
-});
-
-//Login functionality
-//check logged in state
-// app.get("/login", (req, res) => {
-//   if (req.session.user) {
-//     res.send({ loggedIn: true, user: req.session.user });
-//   } else {
-//     res.send({ loggedIn: false });
-//   }
-// });
-
-// app.post("/logout", (req, res) => {
-//   req.session.destroy();
-//   req.session = null;
-//   res.send("User logged out");
-// });
-
 //create JWT aauth
 const verifyJWT = (req, res, next) => {
   const token = req.headers["x-access-token"];
@@ -249,49 +165,6 @@ const verifyJWT = (req, res, next) => {
 //check authentication
 app.get("/isUserAuth", verifyJWT, (req, res) => {
   res.send("Authenticated");
-});
-
-app.post("/login", (req, res) => {
-  const userName = req.body.userName;
-  const password = req.body.password;
-
-  db.query(
-    "SELECT * FROM users WHERE userName = ?",
-    userName,
-    (err, result) => {
-      if (err) {
-        res.send({ err: err });
-      }
-
-      if (result.length > 0) {
-        bcrypt.compare(password, result[0].password, (err, response) => {
-          if (response) {
-            //create session
-            req.session.user = result;
-
-            const id = result[0].id;
-
-            //replace jwtsecret with env file
-            const token = jwt.sign({ id }, "JWTSECRET", {
-              expiresIn: 300,
-            });
-
-            req.session.user = result;
-
-            console.log(req.session.user);
-            res.json({ auth: true, token: token, result: result });
-          } else {
-            res.send({
-              auth: false,
-              message: "Wrong username / password combination",
-            });
-          }
-        });
-      } else {
-        res.json({ auth: false, message: "No user Exists" });
-      }
-    }
-  );
 });
 
 //Create sell functionality
