+++ server/s3.js
@@ -1,30 +1,27 @@
-const S3 = require('aws-sdk/clients/s3');
+const S3 = require("aws-sdk/clients/s3");
 const multer = require("multer");
-const fs = require('fs');
-require('dotenv').config();
-
+const fs = require("fs");
+require("dotenv").config();
 
 const bucketName = process.env.AWS_BUCKET_NAME;
 const region = process.env.AWS_BUCKET_REGION;
 const accessKeyId = process.env.AWS_ACCESS_KEY;
 const secretAccessKey = process.env.AWS_SECRET_KEY;
 
-
 const s3 = new S3({
-    region,
-    accessKeyId,
-    secretAccessKey
-  });
+  region,
+  accessKeyId,
+  secretAccessKey,
+});
 
- function uploadFile(file) {
-     
-    const fileStream = fs.createReadStream(file.path);
-  
-    const uploadParams = {
-      Bucket: bucketName,
-      Body: fileStream,
-      Key: file.filename,
-    }
-    return s3.upload(uploadParams).promise();
-  }
-  exports.uploadFile = uploadFile;
+function uploadFile(file) {
+  const fileStream = fs.createReadStream(file.path);
+
+  const uploadParams = {
+    Bucket: bucketName,
+    Body: fileStream,
+    Key: file.filename,
+  };
+  return s3.upload(uploadParams).promise();
+}
+exports.uploadFile = uploadFile;
